Change design to use Step class inheritance for "session" variables Begin login step separation
