+++ ui_tests/Include/features/output_checker/adjudicate_request.feature
@@ -3,7 +3,8 @@ Feature: Adjudicate request
   As an output checker I need the ability to change the status of a request to "Approved" so that I can allow the requester to take their outputs out of the SRE
 
 	Background:
-    Given an unclaimed request exists
+	  Given requester has logged in
+    And requester has submitted a request
     And output checker has logged in
     When output checker tries to claim an unclaimed request
 
@@ -14,4 +15,5 @@ Feature: Adjudicate request
   Scenario: Request is worthy of approval
     And the output checker marks the request as approved
     Then the output checker should see the status of the request updated to 'Approved'
+    And requester has logged into download interface
     And the approved files are available for download outside of the secure environment
