--- ui_tests/Include/scripts/groovy/test/ocwa/common/LoginStep.groovy
@@ -1,4 +1,0 @@
-package test.ocwa.common
-
-class LoginStep {
-}
