+++ ui_tests/Include/scripts/groovy/test/ocwa/common/Step.groovy
@@ -1,0 +1,10 @@
+package test.ocwa.common
+
+/**
+ * Base Step class used to store "session" variables across the lifecycle of a specific test run
+ * @author Jeremy Ho
+ */
+public class Step {
+	// Edit these variables only when necessary in children classes
+	protected static def G_REQUESTNAME = ''
+}
