+++ ui_tests/Libs/CustomKeywords.groovy
@@ -3,7 +3,3 @@
  * This class is generated automatically by Katalon Studio and should not be modified or deleted.
  */
 
-
-def static "test.ocwa.common.Utils.gen_random_test_request_name"() {
-    (new test.ocwa.common.Utils()).gen_random_test_request_name()
-}
